+++ src/main/java/com/imsweb/mph/mpgroups/Mp2007BreastGroup.java
@@ -50,7 +50,7 @@ public class Mp2007BreastGroup extends MphGroup {
                 TempRuleResult result = new TempRuleResult();
                 if (!GroupUtility.validPairedSiteLaterality(i1.getLaterality(), i2.getLaterality())) {
                     result.setPotentialResult(MphUtils.MpResult.MULTIPLE_PRIMARIES);
-                    result.setMessage("Unable to apply Rule " + this.getStep() + " of " + this.getGroupId() + ". Valid and known laterality should be provided.");
+                    result.setMessageUnknownLaterality(this.getStep(), this.getGroupId());
                 }
                 else if (GroupUtility.areOppositeSides(i1.getLaterality(), i2.getLaterality()))
                     result.setFinalResult(MphUtils.MpResult.MULTIPLE_PRIMARIES);