+++ src/test/java/com/imsweb/mph/HematoDbLab.java
@@ -13,9 +13,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.opencsv.CSVReaderBuilder;
 import com.opencsv.CSVWriter;
-import com.opencsv.CSVWriterBuilder;
 
 import com.imsweb.seerapi.client.NotFoundException;
 import com.imsweb.seerapi.client.SeerApi;
@@ -29,10 +27,11 @@ public class HematoDbLab {
         createHematoDbCsvFiles();
     }
 
+    @SuppressWarnings("ConstantConditions")
     private static void createHematoDbCsvFiles() throws IOException {
-        File samePrimaryFile = new File(System.getProperty("user.dir") + "/src/main/resources/Hematopoietic2010SamePrimaryPairs.csv");
-        File transformToFile = new File(System.getProperty("user.dir") + "/src/main/resources/Hematopoietic2010TransformToPairs.csv");
-        File transformFromFile = new File(System.getProperty("user.dir") + "/src/main/resources/Hematopoietic2010TransformFromPairs.csv");
+        File samePrimaryFile = new File(getWorkingDirectory() + "/src/main/resources/Hematopoietic2010SamePrimaryPairs.csv");
+        File transformToFile = new File(getWorkingDirectory() + "/src/main/resources/Hematopoietic2010TransformToPairs.csv");
+        File transformFromFile = new File(getWorkingDirectory() + "/src/main/resources/Hematopoietic2010TransformFromPairs.csv");
 
         try (CSVWriter samePrimaryWriter = new CSVWriter(new OutputStreamWriter(new FileOutputStream(samePrimaryFile), StandardCharsets.UTF_8));
              CSVWriter transformToWriter = new CSVWriter(new OutputStreamWriter(new FileOutputStream(transformToFile), StandardCharsets.UTF_8));
@@ -61,13 +60,21 @@ public class HematoDbLab {
                 offset += 100;
             } while (allDiseases.size() < total);
 
+            // Get the complete values of each of the diseases.
+            Map<String, Disease> allFullDiseases = new HashMap<>();
             if (total == allDiseases.size()) {
+                for (Disease d : allDiseases) {
+                    Disease disease = api.disease().getById("latest", d.getId()).execute().body();
+                    allFullDiseases.put(disease.getId(), disease);
+                }
+            }
+
+            if (allFullDiseases.size() > 0) {
                 List<String[]> samePrimaryPairs = new ArrayList<>(), transformTo = new ArrayList<>(), transformFrom = new ArrayList<>();
                 samePrimaryPairs.add(new String[] {"morphology", "start year", "end year", "same primary"});
                 transformTo.add(new String[] {"morphology", "start year", "end year", "transform to"});
                 transformFrom.add(new String[] {"morphology", "start year", "end year", "transform from"});
-                for (Disease d : allDiseases) {
-                    Disease disease = api.disease().getById("latest", d.getId()).execute().body();
+                for (Disease disease : allFullDiseases.values()) {
                     String morphology = disease.getIcdO3Morphology();
                     if (disease.getSamePrimaries() != null)
                         for (YearRangeString range : disease.getSamePrimaries()) {
@@ -78,7 +85,7 @@ public class HematoDbLab {
                                     range.getEndYear() != null ? range.getEndYear() : (disease.getValid() != null && disease.getValid().getEndYear() != null ? disease.getValid().getEndYear() : 9999);
                             Disease samePrimary = null;
                             try {
-                                samePrimary = api.disease().getById("latest", range.getValue()).execute().body();
+                                samePrimary = allFullDiseases.get(range.getValue());
                             }
                             catch (NotFoundException e) {
                                 //If a disease is listed as same primary but not existed, don't add it.
@@ -96,7 +103,7 @@ public class HematoDbLab {
                                     range.getEndYear() != null ? range.getEndYear() : (disease.getValid() != null && disease.getValid().getEndYear() != null ? disease.getValid().getEndYear() : 9999);
                             Disease transformToMorphology = null;
                             try {
-                                transformToMorphology = api.disease().getById("latest", range.getValue()).execute().body();
+                                transformToMorphology = allFullDiseases.get(range.getValue());
                             }
                             catch (NotFoundException e) {
                                 //If a disease is listed as transform to but not existed, don't add it.
@@ -114,7 +121,7 @@ public class HematoDbLab {
                                     range.getEndYear() != null ? range.getEndYear() : (disease.getValid() != null && disease.getValid().getEndYear() != null ? disease.getValid().getEndYear() : 9999);
                             Disease transformFromMorphology = null;
                             try {
-                                transformFromMorphology = api.disease().getById("latest", range.getValue()).execute().body();
+                                transformFromMorphology = allFullDiseases.get(range.getValue());
                             }
                             catch (NotFoundException e) {
                                 //If a disease is listed as transform from but not existed, don't add it.
@@ -130,5 +137,9 @@ public class HematoDbLab {
             else
                 System.out.println("Something wasn't right. The total number of diseases you got is different from what the API returns. Please try again.");
         }
+    }
+
+    public static String getWorkingDirectory() {
+        return System.getProperty("user.dir").replace(".idea\\modules", ""); // this will make it work in IntelliJ and outside of it...
     }
 }